Improve : 코드래빗 리뷰 반영
