+++ DataAccess/UmpaApi/ResponseModels/GetChatRoomResponse.swift
@@ -1,0 +1,5 @@
+// Created for Umpa in 2025
+
+import Foundation
+
+struct GetChatRoomResponse: Decodable {}
