+++ Umpa/Application/UmpaApp.swift
@@ -15,9 +15,8 @@ struct UmpaApp: App {
     @InjectedObject(\.appState) private var appState
     @Injected(\.appInteractor) private var appInteractor
 
-    let cancelBag = CancelBag()
-
     init() {
+        NetworkMonitor.shared.start()
         prepareKakaoLogin()
         prepareNaverLogin()
     }
@@ -47,9 +46,6 @@ struct UmpaApp: App {
                     } catch {
                         UmpaLogger.log("Error restoring sign-in: \(error)", level: .error)
                     }
-                }
-                .onFirstAppear {
-                    NetworkMonitor.start()
                 }
         }
     }
